refactor: move classes to new project structure
Move the service classes into a service package outside of routing. This is so they can more
cleanly be reused when multiple routes start using them.
