+++ src/main/kotlin/com/chesire/benzaiten/plugins/koin/modules/ProfileModule.kt
@@ -1,6 +1,6 @@
 package com.chesire.benzaiten.plugins.koin.modules
 
-import com.chesire.benzaiten.routing.profile.service.ProfileService
+import com.chesire.benzaiten.service.profile.ProfileService
 import org.koin.dsl.module
 
 val profileServiceModule = module {
