+++ src/main/kotlin/com/chesire/benzaiten/routing/auth/Auth.kt
@@ -1,6 +1,6 @@
 package com.chesire.benzaiten.routing.auth
 
-import com.chesire.benzaiten.routing.auth.service.AuthService
+import com.chesire.benzaiten.service.auth.AuthService
 import io.ktor.server.application.call
 import io.ktor.server.response.respondText
 import io.ktor.server.routing.Route
@@ -11,6 +11,7 @@ import io.ktor.server.routing.get
  */
 fun Route.auth() {
     val authService = AuthService()
+
     get("auth/") {
         authService.performRequest()
     }
