+++ src/main/kotlin/com/chesire/benzaiten/service/profile/ProfileService.kt
@@ -1,0 +1,24 @@
+package com.chesire.benzaiten.service.profile
+
+import com.chesire.benzaiten.ext.toResult
+import com.chesire.benzaiten.service.SpotifyErrorDto
+import com.github.michaelbull.result.Result
+import io.ktor.client.HttpClient
+import io.ktor.client.request.accept
+import io.ktor.client.request.bearerAuth
+import io.ktor.client.request.get
+import io.ktor.http.ContentType
+
+private const val PROFILE_URL = "https://api.spotify.com/v1/me"
+
+class ProfileService(private val httpClient: HttpClient) {
+
+    suspend fun retrieveUserProfile(bearerToken: String): Result<SpotifyProfileDto, SpotifyErrorDto> {
+        return httpClient
+            .get(PROFILE_URL) {
+                accept(ContentType.Application.Json)
+                bearerAuth(bearerToken)
+            }
+            .toResult()
+    }
+}
